Add splash component
