+++ package.json
@@ -18,7 +18,8 @@
     "react-intl": "^6.5.1",
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.17.0",
-    "redux-logger": "^3.0.6"
+    "redux-logger": "^3.0.6",
+    "spin.js": "^4.1.1"
   },
   "devDependencies": {
     "@types/node": "^20.8.8",
