+++ src/App.css
@@ -7,4 +7,5 @@
 @import "./styles/_inputs.scss";
 @import "./styles/_reset.scss";
 @import "./styles/_tables.scss";
+@import "./styles/_splash.scss";
 @import "./styles/_typography.scss";
