+++ src/apis/scim.ts
@@ -4,7 +4,8 @@ import { AppDispatch, AppRootState } from "init-app";
 export const baseURL = "https://api.eduid.docker/scim/";
 
 export const accessTokenTest =
-  "eyJhbGciOiJFUzI1NiJ9.eyJhdWQiOiJlZHVpZC5kb2NrZXIiLCJhdXRoX3NvdXJjZSI6ImNvbmZpZyIsImV4cCI6MTY5OTU0MjM1NSwiaWF0IjoxNjk5NTM4NzU1LCJpc3MiOiJhcGkuZWR1aWQuZG9ja2VyIiwibmJmIjoxNjk5NTM4NzU1LCJyZXF1ZXN0ZWRfYWNjZXNzIjpbeyJzY29wZSI6ImVkdWlkLnNlIiwidHlwZSI6InNjaW0tYXBpIn1dLCJzY29wZXMiOlsiZWR1aWQuc2UiXSwic291cmNlIjoiY29uZmlnIiwic3ViIjoiZWR1aWRfbWFuYWdlZF9hY2NvdW50c18xIiwidmVyc2lvbiI6MX0.hfD_S8RV2ANdS4VtK9sb2FzGH3HVLKrUZkuWi7NJFz4niznrMtyRYHCX4Z6JzUzPtzRxs-ov-YUHOtmolz3rlA";
+  "eyJhbGciOiJFUzI1NiJ9.eyJhdWQiOiJlZHVpZC5kb2NrZXIiLCJhdXRoX3NvdXJjZSI6ImNvbmZpZyIsImV4cCI6MTY5OTYwNzc3OSwiaWF0IjoxNjk5NjA0MTc5LCJpc3MiOiJhcGkuZWR1aWQuZG9ja2VyIiwibmJmIjoxNjk5NjA0MTc5LCJyZXF1ZXN0ZWRfYWNjZXNzIjpbeyJzY29wZSI6ImVkdWlkLnNlIiwidHlwZSI6InNjaW0tYXBpIn1dLCJzY29wZXMiOlsiZWR1aWQuc2UiXSwic291cmNlIjoiY29uZmlnIiwic3ViIjoiZWR1aWRfbWFuYWdlZF9hY2NvdW50c18xIiwidmVyc2lvbiI6MX0.WNRNPAy9Sm_DbKDAE-tS3SJWyYAevZagJo7AybS6GtJZNazhYF0r_zGk6-2J1zkblo3yWkALBOC1bDqRMnJ7mA";
+
 const scimHeaders = (token: string) => {
   return {
     "Content-Type": "application/scim+json",
