+++ src/styles/_splash.scss
@@ -1,0 +1,20 @@
+#eduid-splash-and-children {
+  position: relative;
+  display: flex;
+  flex: 1;
+  flex-direction: column;
+}
+
+#eduid-splash-spinner {
+  height: 100%;
+  width: 100%;
+  background-color: $body-gray;
+  position: absolute;
+  z-index: 100;
+
+  .spinner {
+    opacity: 1;
+
+    // top: 40vh !important;
+  }
+}
